R6 class for Simple SBM sampler

R6 class for Simple SBM sampler

R6 class for Simple SBM sampler

Super classes

GSBM::SBM -> GSBM::SBM_sampler -> SimpleSBM_sampler

Active bindings

nbNodes

number of nodes

nbBlocks

number of blocks

nbDyads

number of dyads (potential edges in the network)

memberships

vector of clustering

indMemberships

matrix for clustering memberships

expectation

expected values of connection under the current model

directed

is the network directed or not

Methods

Public methods

SimpleSBM_sampler$new()

SimpleSBM_sampler$rMemberships()

SimpleSBM_sampler$rAdjacency()

SimpleSBM_sampler$show()

SimpleSBM_sampler$clone()

Inherited methods

GSBM::SBM$plot()

GSBM::SBM$print()

Method new()

constructor for SBM

Usage

SimpleSBM_sampler$new(
  model,
  nbNodes,
  directed,
  blockProp,
  connectParam,
  covarParam = numeric(0),
  covarList = list()
)

Arguments

model

character describing the type of model

nbNodes

number of nodes in the network

directed

logical, directed network or not.

blockProp

parameters for block proportions (vector of list of vectors)

connectParam

list of parameters for connectivity with a matrix of means 'mu' and an optional scalar for the variance 'sigma2'. The size of mu must match blockProp length

covarParam

optional vector of covariates effect

covarList

optional list of covariates data

Method rMemberships()

a method to generate a vector of block indicators

Usage

SimpleSBM_sampler$rMemberships()

Returns

nothing (sampled memberships is stored in the current object)

Method rAdjacency()

a method to sample an adjacency matrix for the current SBM

Usage

SimpleSBM_sampler$rAdjacency()

Returns

nothing (sampled adjacency matrix is stored in the current object)

Method show()

show method

Usage

SimpleSBM_sampler$show(type = "Sampler for a Simple Stochastic Block Model")

Arguments

type

character used to specify the type of SBM

Method clone()

The objects of this class are cloneable with this method.

Usage

SimpleSBM_sampler$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.